api-2/requirements.txt
Flask
Flask-Cors
pandas
numpy
scikit-learn
mysql-connector-python
sqlalchemy
